+++ backend-core/src/main/java/com/promptpicture/backend/core/cart/domain/Cart.java
@@ -18,5 +18,5 @@ public class Cart {
     private BigDecimal totalPriceWithVat;
     private BigDecimal vatRate;
     private List<CartItem> cartItemOutputList;
-    private OffsetDateTime updateAt;
+    private OffsetDateTime updatedAt;
 }