+++ backend-core/src/main/java/com/promptpicture/backend/core/prompt/adapter/PromptRepositoryAdapter.java
@@ -20,4 +20,6 @@ public interface PromptRepositoryAdapter {
 
     void deleteByPromptId(Long promptId);
 
+    List<Prompt> getListOfAllPrompts();
+
 }