+++ entrypoints/quartz/src/main/java/com/promptpicture/backend/entrypoint/quartz/prompt/common/ClearPromptHistoryJobIdentifiers.java
@@ -1,0 +1,12 @@
+package com.promptpicture.backend.entrypoint.quartz.prompt.common;
+
+import lombok.experimental.UtilityClass;
+
+@UtilityClass
+public class ClearPromptHistoryJobIdentifiers {
+
+    public static final String CLEAR_PROMPT_HISTORY = "CLEAR_PROMPT_HISTORY";
+    public static final String CLEAR_PROMPT_HISTORY_JOB = "clearPromptHistoryJob";
+    public static final String CLEAR_PROMPT_HISTORY_TRIGGER= "clearPromptHistoryTrigger";
+
+}