+++ changelog.md
@@ -11,6 +11,7 @@
 - aitest-24 - add get prompt detail
 - aitest-33 - add price table
 - aitest-31 - add job for removing cart older than 2 days
+- aitest-5 - add job for removing prompt older than 12 hours
 
 ### Fix